requirements.txt
flask==3.0.0
anthropic==0.40.0
gspread==6.1.2
google-auth==2.28.0
